Separate base app
Other changes:
- Add login template (not responsive yet)
