+++ baseapp/static/css/base.css
@@ -1,0 +1,72 @@
+:root {
+    --heading-font: 'Roboto Slab', sans-serif;
+    --main-font: 'Exo 2', sans-serif;
+    --color-1: #1F363D;
+    --color-1-light: #27434C;
+    --color-1-dark: #17282D;
+    --color-2: #40798C;
+    --color-2-light: #47869B;
+    --color-2-dark: #396C7C;
+    --color-3: #70A9A1;
+    --color-3-light: #79B7AF;
+    --color-3-dark: #649992;
+    --color-4: #9EC1A3;
+    --color-4-light: #ABD1B1;
+    --color-4-dark: #92B297;
+    --color-5: #CFE0C3;
+    --color-5-light: #DDEFD0;
+    --color-5-dark: #C1D1B5;
+}
+
+html * {
+    font-family: var(--main-font);
+}
+
+h1,
+h2 {
+    font-family: var(--heading-font);
+}
+
+.md-btn {
+    font-size: 1rem;
+    display: inline-block;
+    height: 36px;
+    min-width: 88px;
+    padding: 6px 16px;
+    line-height: 1.42857143;
+    text-align: center;
+    white-space: nowrap;
+    vertical-align: middle;
+    cursor: pointer;
+    border: 0;
+    border-radius: .25rem;
+    outline: 0;
+    box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 1px 5px 0 rgba(0, 0, 0, 0.12), 0 3px 1px -2px rgba(0, 0, 0, 0.2);
+    transition: box-shadow 0.28s cubic-bezier(0.4, 0, 0.2, 1), background-color 0.28s cubic-bezier(0.4, 0, 0.2, 1);
+    overflow: hidden;
+    position: relative;
+}
+
+.md-btn:hover {
+    box-shadow: 0 8px 10px 1px rgba(0, 0, 0, 0.14), 0 3px 14px 2px rgba(0, 0, 0, 0.12), 0 5px 5px -3px rgba(0, 0, 0, 0.4);
+}
+
+/* button colors */
+
+.md-btn-primary {
+    background: var(--color-1);
+    color: #fff;
+}
+
+.md-btn-primary:focus, .md-btn-primary:hover {
+    background: var(--color-1-light)
+}
+
+.md-btn-secondary {
+    background: white;
+    color: var(--color-2);
+}
+
+.md-btn-secondary:focus, .md-btn-secondary:hover {
+    background: #DDD;
+}
