+++ baseapp/templates/core/index.html
@@ -1,0 +1,16 @@
+{% extends "core/base.html" %}
+{% load static %}
+{% block title %}
+    {{ "Welcome "}}
+{% endblock title %}
+
+{% block styles %}
+    <link rel="stylesheet" href="{% static 'css/index.css' %}">
+{% endblock styles %}
+
+{% block content %}
+    <div class="container index-container">
+        <h1 class="">TECHNOSTALL MANAGER</h1>
+        <a class="" href="{% url 'login' %}">Login</a>
+    </div>
+{% endblock content %}
