+++ baseapp/templates/core/login.html
@@ -1,0 +1,67 @@
+{% extends "core/base.html" %}
+{% load static %}
+{% block title %}
+    {{ "Login "}}
+{% endblock title %}
+
+{% block styles %}
+    <link rel="stylesheet" href="{% static 'css/login.css' %}">
+{% endblock styles %}
+
+{% block content %}
+<div class="container login-container">
+    <div class="login-form-container">
+        <div class="login-form-header">
+            <h1>TECHNOSTALL MANAGER</h1>
+        </div>
+        <form class="login-form" method="POST" action="{% url 'login' %}">
+            <h2>Login</h2>
+            {% csrf_token %}
+            {% if next %}
+                {% if user.is_authenticated %}
+                <small class="text-danger">Your account doesn't have access to this page. To proceed,
+                please login with an account that has access.</small>
+                {% else %}
+                <small class="text-danger">Please login to see this page.</small>
+                {% endif %}
+            {% endif %}
+            <div class="form-group">
+                <label for="id_username">Username</label>
+                <input type="text"
+                    class="form-control" name="username" id="id_username" aria-describedby="help_username" placeholder="Your username...">
+                <small id="help_username" class="form-text text-muted">Enter your username</small>
+            </div>
+            <div class="form-group">
+                <label for="id_password">Password</label>
+                <input type="password" class="form-control" name="password" id="id_password" placeholder="">
+            </div>
+            {% if form.errors %}
+            <small class="text-danger">Your username and password didn't match. Please try again.</small>
+            {% endif %}
+            <input class="md-btn md-btn-primary" type="submit" value="login" />
+            <input type="hidden" name="next" value="{{ next }}" />
+        </form>
+    </div>
+</div>
+
+<div style="display: none">
+        
+        <form method="post" action="{% url 'login' %}">
+        {% csrf_token %}
+        <table>
+        <tr>
+            <td>{{ form.username.label_tag }}</td>
+            <td>{{ form.username }}</td>
+        </tr>
+        <tr>
+            <td>{{ form.password.label_tag }}</td>
+            <td>{{ form.password }}</td>
+        </tr>
+        </table>
+        
+        <input type="submit" value="login" />
+        <input type="hidden" name="next" value="{{ next }}" />
+        </form>
+</div>
+
+{% endblock content %}
