+++ baseapp/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from django.contrib.auth import views as auth_views
+from . import views
+
+urlpatterns = [
+    path('login/', auth_views.LoginView.as_view(template_name='core/login.html'), name='login'),
+    path('', views.index, name='index')
+]
