+++ baseapp/views.py
@@ -1,0 +1,6 @@
+from django.shortcuts import render
+from django.http import HttpRequest
+
+# Create your views here.
+def index(request):
+    return render(request, 'core/index.html')
