--- itemmanager/static/css/base.css
@@ -1,7 +1,0 @@
-h1 {
-    font-family: 'Roboto Slab', sans-serif;
-}
-
-p, a {
-    font-family: 'Exo 2', sans-serif
-}
