+++ itemmanager/views.py
@@ -3,5 +3,4 @@ from django.http import HttpResponse
 
 
 # Create your views here
-def index(request):
-    return render(request, 'core/index.html')
+
