+++ technostall/settings.py
@@ -37,6 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'baseapp',
     'itemmanager',
     'usermanager'
 ]
@@ -56,7 +57,9 @@ ROOT_URLCONF = 'technostall.urls'
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [],
+        'DIRS': [
+            'technostall/templates',
+        ],
         'APP_DIRS': True,
         'OPTIONS': {
             'context_processors': [
@@ -118,6 +121,10 @@ USE_TZ = True
 
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/2.0/howto/static-files/
+
+# STATICFILES_DIRS = [
+#     os.path.join(BASE_DIR, "technostall/static")
+# ]
 
 STATIC_URL = '/static/'
